Unique CD/DVD identification function written
Comparison based on track length, or track start position, is a very
reliable metric. It is also 'portable', as we can assume any CD or DVD
parsing library would have either of those pieces of functionality. And
if it has one, but not the other, it's easily derived.

Switching back to an assumed run as root. Was running into too many
issues otherwise. Permissions, access, etc.